backend/requirements.txt
a2wsgi==1.10.10
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioitertools==0.11.0
aiokafka>=0.10.0
aiosignal==1.3.2
aiosmtplib==5.0.0
aiosqlite==0.21.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
apache-airflow==3.1.3
apache-airflow-core==3.1.3
apache-airflow-providers-common-compat==1.9.0
apache-airflow-providers-common-io==1.6.5
apache-airflow-providers-common-sql==1.29.0
apache-airflow-providers-http==5.3.0
apache-airflow-providers-imap==3.9.0
apache-airflow-providers-postgres==6.2.0
apache-airflow-providers-standard==1.9.2
apache-airflow-task-sdk==1.1.3
apispec==6.8.2
appnope==0.1.4
argcomplete==3.6.3
asgiref==3.11.0
asttokens==3.0.0
asyncio-throttle>=1.0.2
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.4.0
babel==2.17.0
backcall==0.2.0
backports.strenum==1.3.1
beautifulsoup4==4.13.4
billiard==4.2.1
bitarray==2.7.4
blinker==1.9.0
boto3==1.26.118
botocore==1.29.118
build==1.3.0
cachelib==0.13.0
cachetools==5.3.3
cadwyn==5.4.5
cattrs==22.2.0
celery==5.5.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
ckzg==2.1.1
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
colorama==0.4.6
colorlog==6.10.1
ConfigUpdater==3.2
connectorx==0.3.3
contourpy==1.3.1
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==46.0.3
curl_cffi==0.12.0
cycler==0.12.1
cytoolz==0.12.1
dataclasses-json==0.6.7
db-dtypes==1.2.0
decorator==5.1.1
Deprecated==1.3.1
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
dune-client==1.2.1
email-validator==2.3.0
eth-account==0.13.6
eth-hash==0.5.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
exceptiongroup==1.3.0
executing==2.1.0
fastapi==0.117.1
fastapi-cli==0.0.16
fastjsonschema==2.21.1
Flask==2.2.5
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-WTF==1.2.2
flower==2.0.1
fonttools==4.55.1
frozendict==2.4.6
frozenlist==1.6.0
gcsfs==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-api-core>=2.21.0
google-api-python-client==2.132.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage>=2.27.0
google-cloud-core==2.4.2
google-cloud-storage==3.1.0
google-crc32c==1.5.0
google-re2==1.1.20240702
google-resumable-media==2.7.2
googleapis-common-protos>=1.72.0
graphql-core==3.2.4
greenback==1.2.1
greenlet==3.2.4
grpcio==1.76.0
grpcio-status>=1.62.2
gunicorn==23.0.0
h11==0.16.0
h2==4.1.0
hexbytes==1.3.0
hpack==4.0.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.7.1
httpx==0.28.1
humanize==4.12.3
hyperframe==6.0.1
idna==3.11
importlib-resources==6.1.1
importlib_metadata==8.7.0
inflection==0.5.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.5.0
jmespath==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
kombu==5.5.3
lazy-object-proxy==1.12.0
libcst==1.8.6
limits==5.2.0
linkify-it-py==2.0.3
lockfile==0.12.2
lru-dict==1.2.0
lxml==5.2.1
Mako==1.3.10
Markdown==3.4.3
MarkupSafe==3.0.3
marshmallow==3.26.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.2.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.8.0
msgspec==0.19.0
multidict==6.4.4
multitasking==0.0.12
mypy-extensions==1.0.0
natsort==8.4.0
ndjson==0.3.1
nextcord==2.6.0
numpy==1.24.2
oauthlib==3.2.2
oli-python==1.2.1
openai==1.42.0
ordered-set==4.1.0
orjson==3.10.7
outcome==1.3.0.post0
packaging==25.0
pandas==2.0.0
pangres==4.1.4
parsimonious==0.10.0
parso==0.8.4
pathspec==0.12.1
peewee==3.18.2
pendulum==3.1.0
pillow==10.2.0
pip-tools==7.5.2
pipdeptree==2.30.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
playwright==1.54.0
pluggy==1.6.0
polars==1.1.0
prison==0.2.1
prometheus_client==0.22.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf>=5.26.0,<6
proto-plus>=1.25.0
psutil==7.1.3
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyairtable==2.3.0.post1
pyarrow==11.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.23
pycryptodome==3.18.0
pydantic==2.12.4
pydantic_core==2.41.5
pyee==13.0.0
PyGithub==2.8.1
Pygments==2.19.2
pygtrie==2.5.0
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.0.9
pyproject_hooks==1.2.0
pyrsistent==0.19.3
PySocks==1.7.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
pytzdata==2020.1
pyunormalize==16.0.0
PyYAML==6.0.3
redis==5.2.1
referencing==0.37.0
regex==2023.6.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retryhttp==1.4.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==14.2.0
rich-argparse==1.7.2
rich-toolkit==0.16.0
rlp==4.1.0
rpds-py==0.29.0
rsa==4.9
s3fs==0.4.2
s3transfer==0.6.0
selenium==4.18.1
setproctitle==1.3.7
sgqlc==16.4
shellingham==1.5.4
simplejson==3.19.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.44
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.0
sqlparse==0.5.3
starlette==0.48.0
StrEnum==0.4.15
structlog==25.5.0
supervisor==4.2.5
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.2.0
text-unidecode==1.3
tinycss2==1.4.0
tomli==2.3.0
toolz==0.12.0
tornado==6.5
tqdm==4.66.5
trio==0.24.0
trio-websocket==0.11.1
typer==0.20.0
types-python-dateutil==2.8.19.13
types-PyYAML==6.0.12.10
types-requests==2.31.0.1
types-urllib3==1.26.25.13
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.20
uuid6==2025.0.1
uvicorn==0.38.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.13
web3==7.10.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==2.0.1
wsproto==1.2.0
WTForms==3.2.1
yarl==1.20.0
yfinance==0.2.65
zipp==3.23.0
